+++ plugins/source/heroku/resources/services/enterprise_account.go
@@ -13,7 +13,7 @@ import (
 func EnterpriseAccounts() *schema.Table {
 	return &schema.Table{
 		Name:        "heroku_enterprise_accounts",
-		Description: `https://devcenter.heroku.com/articles/platform-api-reference#enterprise-account-attributes`,
+		Description: `https://devcenter.heroku.com/articles/platform-api-reference#enterprise-account`,
 		Resolver:    fetchEnterpriseAccounts,
 		Columns: []schema.Column{
 			{